ENRT.LinuxPerfMeasurementGenerator: fix generator return implementation
When 'yield' is used in python it automatically changes the
function/method into a generator. This is regardless of if the yield is
reached or not - this is decided when the function is parsed, not during
runtime.

At the same time, the 'return' statement works differently in a
generator. The return statement raises the 'StopIteration' exception
that is used by loops etc. to detect that the generation has stopped. If
'return' is used with a value, this value is provided to the exception
object:

    return val is sort of equivalent to raise StopIteration(val)

However, when we write:

    for i in generator:

The `StopIteration(val)` simply exits the loop, the `val` itself isn't
assigned to `i` to start one more iteration of the loop.

Because of this the old code of the LinuxPerfMeasurementGenerator when
it simply called `return super().generate_perf_measurements_combinations`
the previous level of the collaborative mixin hierarchy stopped
iteration and didn't receive the previously generated list of
measurement combinations.

To fix this I restructured the method to avoid the use of the `return`
statement, and instead relied on the use of `yield` for the individual
previous values.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>