ENRT.DevInterruptHWConfigMixin: fix grep regex when bus_info missing
When bus_info is missing, e.g. for container interfaces, the generated
regex is:

    '(eth0)|()'

The '()' part matches anything and so the original output of the
/proc/interrupts file is returned fully, then the '| cut | sed' reduces
it to the first column, but the header also stays.

This all leads to a mangled unexpected output of the command that is
then simply passed to the 'strip().split()' with an attempt to convert
to integers.

The solution is to not generate a regex that matches anything - so skip
the bus_info part.

Secondly, in case the interfaces itself doesn't have any interrupts
present in the /proc/interrupts file, the final list comprehension needs
to be updated to ignore empty strings to simply generate an empty list.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>