RecipeCommon.Perf.BaselineEvaluator: add data to result object
Having access to the evaluation data as part of the created final Result
object can be very helpful for post run analysis and is difficult to
find without this direct link.

As the data is already present somewhere in the lrc file this shouldn't
add any overhead except a simple object link to one additional place.

This also includes a minor renaming of the variables involved to clear
up any confusions:
* comparison_result -> cumulative result is what contains the merged
  value of all baseline comparisons that we report at the end
* comparison -> comparison_result is what contains the result of a
  single baselne comparison that we do and is added to the cumulative
  result based on max_severity

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>