LinuxPerfMeasurementGenerator: add linuxperf_cpus_override parameter
This is to override the linuxperf_cpus property defined by a recipe that
uses the LinuxPerfMeasurementGenerator extension.

In following example two groups of cpus will be profiled separately:

linuxperf_cpus_override=[ [0,1], [2,3] ]

This patch explicitly disallows use of empty list of cpus to be
profiled. In that case, an exception is raised.

Possibility to profile the whole system (an empty list of cpus) will be
added later in future.

Signed-off-by: Jan Tluka <jtluka@redhat.com>