add .dockerignore
When the podman container lnst-agent image is build, the content of the
directory is copied to the image along with .venv that may be present.

This however may contain incorrect path to the python executable.

To avoid this, the .dockerignore is added that excludes the .venv
directory from COPY directive in Dockerfile.

Signed-off-by: Jan Tluka <jtluka@redhat.com>