RecipeCommon.Perf.BaselineEvaluator: refactor compare_result_with_baseline
The main point of this refactoring is to split out the return value of
the `compare_result_with_baseline` method so that it returns a list of
individual metric comparisons instead of a single merged comparison of
all evaluated metrics.

As single metric comparison is described by multiple values I introduce
a new dataclass `MetricComparison` that defines a type that should make
it easier to work with.

This means that the return value of the compare_result_with_baseline
method changes to `List[MetricComparison]`.

For BaselineFlowAverageEvaluator we evaluate the list of metrics in the
`_metrics_to_evaluate` list.

For BaselineCPUAverageEvaluator we evaluate the `utilization` metric. As
we have now added the result index to the metric and threshold name
there is no more a need to do a rename to `host.id` here.

The base class `BaselineEvaluator` now simply raises a
NotImplementedError.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>