Controller.Host: pass params for new NetNamespace
Problem statement:
Once tester creates new NetNamespace, he is unable
to access the params of the run, as they are
initialized and handled by the Host class, which
was always used for the catalog baseline matching.

Fix:
As Host is responsible for the creation and
initialization of the new NetNamespace, it should
also handle the passing of the params to the new
NetNamespace, as it can be used as flow endpoint
to emulate container from the network perspective.

This is a usecase we never had before (having
flows to/from network namespace, different from the
initns), thus we did not see this issue.

However with this fix, we won't be able to tell
the difference between the NetNamespace and Host
baseline at the moment.