OffloadSubConfigMixin: fix case when offload_settings is empty
If a recipe uses this mixin, it must specify at least one offload
configuration to function properly. If empty list is passed, the
performance tests will be skipped.

In some situations, the user may want to run the performance tests with the
default NIC offload settings. This can be done by specifying empty list
and handling the case so that the offload configuration is skipped.

v2:
fix bypass of collaborative inheritance

Signed-off-by: Jan Tluka <jtluka@redhat.com>