docs/source/conf.py: add psutil to mock imports
After the recent change of dependencies we've started using `psutil`
module in our lnst.Common.Utils module. This leads to an undefined
import when building the docs with just the `docs/requirements.txt`
dependencies and creates a lot of empty pages.

Adding the dependency to mock imports fixes the issue as we don't
actually need this when building the documentation.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>