LinuxPerfMeasurement: Use two lists for storing jobs
Previously, the `_jobs` list was initialized once per recipe run and
appended to every perf iteration. The list should be empty at the start
of every iteration.

Splitting the list into two and moving the full job list in the
`finished()` method ensures that at the start of the `start()` method,
the first list is empty.