Refactor Traffic Control result processing
Uses PerfParallelResult for the results of all the tc jobs that are run.

Also, previously every instance of tc was run as a seperate job. Now TrafficControlRunner has been refactored to take a list of batchfiles and uses `asyncio.subprocess` to run multiple instances of tc. This is to better ensure that they are actually run in parallel.