Recipes/ENRT/FlowMeasurementGenerator: allow multiple cpupin strategies
This moves most of the functionality of FlowMeasurementGenerator to base
class BaseFlowMeasurementGenerator.

The base class adds generator_cpupin and receiver_cpupin methods that are
implemented by derived classes and return lists with CPUs that the base class
uses for the generator and receiver CPU pinning based on the user defined policy.

The FlowMeasurementGenerator original behaviour does not change.

v2:
- split cpupin to generator/receiver cpupin methods
- add docstring for _cpupin_based_on_policy()

Signed-off-by: Jan Tluka <jtluka@redhat.com>