RecipeCommon.Perf.MeasurementResults: refactor result description generation
Previously the descriptions were generated directly or almost directly
as part of the Measurement classmethod "report_results" which also
creates a RecipeResult object that is added to the current RecipeRun
instance.

This meant that it wasn't possible to do custom descriptions or custom
result objects without calling this "report_results" method which would
also create the recipe result object.

This refactors the code a bit so that the MeasurementResult description
is generated directly by the MeasurementResult object and the
"report_results" method simply calls this when needed.

Unfortunatelly this means that the format for the CPUMeasurementResults
has to change a bit as these are not grouped by "host" anymore and each
individual result will contain the host id.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>