Agent/Job.py: fix kill() result reporting
On the Agent, there are always 2 instances of a Job class:
* the main process instance - this is the one that the Agent server
  process is using and has the "parent pipe" to read data from:
* the forked process instance - this is the one that actually runs the
  job process and on finishing the job sends result data over "child
  pipe" back to the agent process.

When the Job.kill() method is called, this is in the context of the
Agent process. In case the signal used is SIGKILL, the forked process
doesn't have a chance to send the result data back to the Agent. Instead
we need to "fake" this from the Agent process.

Previously the code called:
1. set_finished
2. send_data

This doesn't work correctly because 'set_finished' closes both the
parent and the child pipes in the Agent process... so even if we attempt
to write into the child pipe, we can't read from it and pass the data
back to the Controller.

Instead the solution is to ONLY call the "send_data" method to write
into the child_pipe.

The Agent process can now properly read from the "parent_pipe" in the
server loop, and when it does so... it also additionally calls the
"set_finished" method which cleans up the opened pipes as well as
forwards the Job results back to the Controller.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>