BaseEnrtRecipe: remove perf_evaluation_strategy parameter
This parameter logically doesn't fit into the generic upstream
BaseEnrtRecipe class.

The reason for this is that from the upstream point of view it doesn't
actually do anything - it only recognized the "none" and "nonzero"
values which... are special use cases and the implementation was a bit
clunky.

We could refactor this to make it make a bit more sense in the upstream
and for it to be extensible - I think this should be an extensible Mixin
class that handles evaluator registration for enrt perf measurements...

But I think it doesn't have a valid use case in upstream yet so I'm
removing it for now. We may want to revisit this later.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>