MPTCPRecipe: Fix disabling rp_filter
rp_filter (reverse path filtering) is responsible for not allowing
responding to packets on a different interface than the packet appeared
on. It has to be disabled for MPTCP on IPv4.

IPv6 doesn't have rp_filter on sysfs, but firewalld or ip6tables might
configure it themselves. We do not run testing on machines where
firewalld is enabled, nor do we use ip6tables so no issues should
appear.