Devices/Device.py: report unsupported feature for switchdev corner case
In some cases a query of switchdev info on a device could return an empty
string. When this happens the code expects either legacy or switchdev
mode in the output, otherwise a DeviceError is raised.

The DeviceError causes termination of the machine's device cleanup
process and some devices may be left without deconfiguration.

Since this case should not be considered as a critical issue a
DeviceFeatureNotSupported is raised instead that is accepted during the
deconfiguration.

Signed-off-by: Jan Tluka <jtluka@redhat.com>