JobContext: Add pop_job() method
Previously, `JobContext.cleanup()` would kill all jobs and clear its
dict. However, the killing of the jobs generates "job_finished" events,
which were previously trying to get the finished job object from job
context's dict and trying to join the job, but the dict was empty at
this point.

By not clearing the job dict and popping the job from the dict in the
"job_finished" handler (instead of just getting it) we shouldn't run
into this problem anymore.