BaseFlowMeasurementGenerator: Do not use client port in mptcp
We are binding the client port to be consistent in multi-flow scenarios.
However, our mptcp scenario opens 2 additional ports and not one. Only
the first one binds to the specified client port, the second one is
random.

This would be fine, but there's a bug where mptcp doesn't close ports
immediately (RHEL-21492). This raises an "Address already in use" error
when the client attempts to bind the client port. We don't bind the
client port to mitigate this issue, but also because, as specified
before, it doesn't matter whether it's bound or not.