Recipes/ENRT/ConfigMixins: Add FirewallMixin.py
The set of classes contained in this module enable users to push
firewall rulesets onto the hosts used for test runs.

There are five flavors to choose from for the actual firewall
implementation, with descriptive names:

* NftablesMixin
* IptablesMixin
* Ip6tablesMixin
* EbtablesMixin
* ArptablesMixin

To use the provided functionality, inherit one of them and either
provide a dictionary of "host: ruleset" entries in a property
'firewall_rulesets' or overwrite the 'firewall_rulesets_generator()'
method with a generator yielding as many dictionaries a desired for
performing a test run with. The latter provides a simple means of
testing with gradually increasing or extending rulesets.

In order to perform post-processing on rulesets after a test run,
override the 'firewall_rulesets' setter method.