PerfMeasurements: add ability to simulate measurements
The "bottom" measurements - meaning the most core measurement in a list
of measurements that we may run can now be "simulated".

This means that instead of actually running the measurement
tools/commands we instead run the following command:

    echo simulated start

And instead of returning real measurement results we return the same
structure of data with a single 0 value perf interval.

Combining this we can now run all of our ENRT recipes in a "simulation"
that preserves the number of commands per test recipe run and should
return results compatible with all of our tooling with just the basic 0
value.

The cpu measurement simulation does actually call the real measurement,
this is because this isn't a "bottom/core" measurement and is instead
designed to always wrap some other measurement.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>