Recipes/SRIOVNetnsOvSRecipe.py: deconfigure ovs bridge before resetting switchdev mode
On a setup with Nvidia CX6DX NIC we hit an issue when switchdev mode of
the NIC could not be set back to legacy mode.

After some testing, the issue was resolved after the switchdev mode was
reset after the ovs bridge device was brought down and ports removed from
the switch. Previously this was done after reset of switchdev.

Logically, the steps of deconfiguration should match the configuration steps,
in reversed order.

Signed-off-by: Jan Tluka <jtluka@redhat.com>