External.TRex.UDPSimple: use src/dst mac and ip addresses
The default trex packet stream generator didn't correctly use the
provided source and destination ip and mac addresses.

This wasn't noticed until now because the test is typically run in a
lab where multiple machine pairs were isolated from eachother.

This however isn't MANDATORY and so to avoid conflicts on a shared
network we should respect the provided values.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>