ENRT/ConfigMixins: replace CoalescingHWConfigMixin with MultiCoalescingHWConfigMixin
The CoalescingHWConfigMixin is only capable of disabling adaptive TX/RX
setting. Typically the user disables adaptive coalescing and configures
specific values of rx/tx-usecs and others. Not setting these would
result in non-deterministic behaviour of the setup.

The MultiCoalescingHWConfigMixin is capable of setting any coalescing
options, so using it instead of CoalescingHWConfigMixin makes sense.

Signed-off-by: Jan Tluka <jtluka@redhat.com>