fix strings with escape sequences
These are used as regex patterns and contain backslashes that python
tries to interpret as an escape sequence.

Starting with python3.12 this is recognized as a:

    SyntaxWarning: invalid escape sequence

This will eventually become a SyntaxError and will break the code.

As such, transforming these into a "raw string" by prepending them with
'r' is the correct solution.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>