BaseTestModule class hierarchy split
BaseTestModule is treated as it was _BaseModule_ when transferring
to agent. So any non-test module would need to inherit from `BaseTestModule`
otherwise it won't be treated as a `ModuleJob` but `ShellJob` instead.

Therefore, this patch includes changes to split class hierarchy,
so now we have `BaseModule` which is supposed to be base class
for all the *Modules and `BaseTestModule` just inherits from it.
So this effectively adds support for non-test modules.