Agent: waiting for condition on agent side
Added support of waiting for condition to be met on agent
side.

This is implemented as a class that inherits from `BaseModule`
to allow more complex conditions with optional context.
The `WaitForConditionModule` is just an abstract class, at least
`condition()` method must be defined. That one is then
called in endless loop once per second until condition
is met or the maximum exection time is exhausted.