Recipes/ENRT: add BondingMixin
This also updates BondRecipe, VlansOverBondRecipe and DoubleBondRecipe with fail_over_mac.

To avoid code duplication in Bond* recipes I moved the common code to new
BondingMixin class.

The class provides the common bonding recipe parameters and a method to create
the bonding devices.

The generate_test_wide_description is added so that it ca be used by the
derived classes.

Signed-off-by: Jan Tluka <jtluka@redhat.com>