XDPBench: exported xdp-bench tool arguments
All the params that are passed to `xdp-bench` tool needs to be
able to use them within LNST.
`XDPBench._prepare_arguments` ignores empty parameters set to
empty string. Thats because not all of the params are applicable
to all XDP bench modes but we still need a way of providing them
to XDPBench so these are passed to XDPBench but set to empty str.

More elegant way would be splitting `XDPBench` into multiple
test modules based on mode it runs...