Recipes.ENRT: always deconfigure
This adjusts the context manager for test_wide_configuration so that the
deconfiguration method is always called.

The deconfiguration is part of the finally block of the context manager,
however the configuration part was called before the try block, this
means that some exceptions raised during the configuration resulted in
the finally block not being called.

This is especially important for problems with LACP switch
deconfiguration.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>