ContainerPoolManager: add custom_lnst network backend
When "custom_lnst" is provided as the plugin to use for networking we
take over the process of connecting the containers manually by creating
our own bridges and veth pairs and moving them to respective containers.

This allows us to run lnst recipes which have machines that are
connected to the same single network multiple times.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>