Recipes/ENRT/GeneveLwtTunnelRecipe.py: adjust _create_perf_flows
This overrides the BaseFlowMeasurementGenerator implementation.

The generator pins iperf processes to the CPUs defined in perf_tool_cpu
list based on the perf_parallel_processes. The GeneveLwtTunnelRecipe
however needs to pin the CPUs based on the flow_count (tunnel flow is
defined by the perf endpoints), otherwise each each of the tunnel flows
would be pinned to the first CPU in the perf_tool_cpu list.

Signed-off-by: Jan Tluka <jtluka@redhat.com>