lnst: Common: ExecCmd: Reduce logging in exec_cmd()
If log_outputs=True (the default), debug log is populated with stdout
and stderr content from the command already. Then if die_on_err=True
(also default) the function raises an exception which callers are
supposed to catch. If useful, they may log from the exception handler.

The side-effect of this is that quite some failing commands are not
logged as errors anymore - callers apparently didn't care to report it
themselves. These are (maybe among others):
- ethtool -a
- devlink dev eswitch show
- cat /sys/class/net/*/device/sriov_numvfs

Signed-off-by: Phil Sutter <psutter@redhat.com>