ENRT.PingFloodRecipe: fix to make the recipe work
Removed the driver parameter, the point of this recipe is to test all
nics on a machine pair so limiting to a single driver is unnecessary.

Retyped the `interval` parameter to Float.

Removed the `mtu` parameter, again, unnecessary for this recipe.

Added a RatePingEvaluator so that the recipe actually reports PASS/FAIL
based on the ping results.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>